api/pyproject.toml
[tool.poetry]
name = "minori-api"
version = "0.dev"
description = "The 'Minori' API"
authors = ["Matteo Corona <matteo.corona@clemps.it>"]
maintainers = ["Roberto Bochet <roberto@demosoftlab.dev>"]
readme = "README.md"
packages = [{ include = "api" }]
include = ["api/font/*.ttf"]

[tool.poetry.dependencies]
python = "^3.9"
flask-pymongo = "^2.3.0"
pathlib = "^1.0.1"
schema = "^0.7.4"
python-dateutil = "^2.8.2"
toml = "^0.10.2"
flask-cors = "^3.0.10"
pillow = "^8.3.2"
pymongo = { extras = ["srv"], version = "^3.12.0" }
flask = "^1.1.4"
werkzeug = "^1.0.1"
jwcrypto = "^1.0.0"
requests = "^2.26.0"
connexion = {version = "^2.9.0", extras = ["swagger-ui"]}
inflection = "^0.5.1"
jsonschema = "3.2.0"
markupsafe = "2.0.1"
gunicorn = "^21.2.0"
python-jwt = "^3.3.4"
fpdf2 = "^2.5.2"
pyyaml = "^6.0.1"
pytz = "^2024.1"
python-dotenv = "^1.0.1"
openai = "^1.29.0"
beautifulsoup4 = "^4.12.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.bandit]
exclude_dirs = [".git", ".mypy_cache", ".pytest_cache", ".tox", "venv", "tests"]
